To summarize the working logic of the project, the script first detects hand gestures from the device's camera and then calculates the midpoint of the palms and the openness of the palms with these points. If your hand is open enough, a mask (the filter itself) is created using the previously provided images and pasted onto the calculated midpoint.
This project was developed to test my computer vision skills. It was created using the Mediapipe and OpenCV libraries, along with the Python programming language. As the name suggests, I attempted to mimic the magic circles that appear on the palms of the popular MARVEL hero Dr. Strange when he casts a spell.

To summarize the working logic of the project, the script first detects hand gestures from the device's camera and then calculates the midpoint of the palms and the openness of the palms with these points. If your hand is open enough, a mask (the filter itself) is created using the previously provided images and pasted onto the calculated midpoint.

Of course, this was a very simplified explanation. There are many things to learn about perspective, rotating image objects, and creating transparent objects during this project. I hope this project will help you understand these topics.